Renaming and moving crash logger. Updating params to log
